default to non forward compatible enums
